sdl: Make widget_id a 32 bit value
See RM #1528

+++ client/gui-sdl2/gui_main.h
@@ -74,7 +74,7 @@ struct theme;
   || ((event).type == SDL_MOUSEBUTTONDOWN                                   \
       && (event).button.button == SDL_BUTTON_LEFT))
 
-typedef Uint16 widget_id;
+typedef Uint32 widget_id;
 
 enum mouse_button_hold_state {
   MB_HOLD_SHORT,
